+++ packages/node/CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased - 2023-06-20
 ### Added
-- Remove `QUERY_ADDRESS_LIMIT` for dictionary queries
+- Add flag `query-address-limit` for dictionary queries
 
 ## [2.6.1] - 2023-06-19
 ### Added